gcp_cloudfunction/requirements.txt
Flask==2.2.5
openai==1.70.0
telethon==1.39.0
python-dotenv==1.1.0
apscheduler==3.11.0